refactor(musea): extract withDefaults parsing to keep native-loader under length limit
